requirements.txt
numpy
opencv-python
Pillow
PyYAML